+++ src/ninja/mbedded/ninjaterm/interfaces/AppendedStringData.java
@@ -1,0 +1,8 @@
+package ninja.mbedded.ninjaterm.interfaces;
+
+/**
+ * Created by gbmhu on 2016-09-22.
+ */
+public interface AppendedStringData {
+    void update(String appendedData);
+}